lab5/builder_tool/pipeline_executor/requirements.txt
google-cloud-aiplatform[pipelines]
argparse